[refac] 유틸리티 관련 또는 범용적으로 쓰이는거 libs 폴더로 전부 관리
